+++ lib/src/json/index.ts
@@ -20,20 +20,43 @@ export function Serialize<TSerialized extends JsonType, TOriginal>(
 }
 
 /**
- * Serialize given serializable type instance to a JSON-compatible type
+ * Serialize given instance to a JSON-compatible value
  * @param serializable Serializable type instance
  * @param options Deflate options
- * @returns JSON-compatible type which can be safely passed to `JSON.serialize`
+ * @returns JSON-compatible value which can be safely passed to `JSON.serialize`
  */
 export function deflate<TOriginal>(
     serializable: TOriginal,
     options?: DeflateOptions<JsonType, TOriginal>
 ): JsonType {
-    return FrontendFunctions(BACKEND_NAME).deflate<JsonType, TOriginal>(serializable, options);
+    const serializedValue = FrontendFunctions(BACKEND_NAME).deflate<JsonType, TOriginal>(serializable, options);
+    if (Util.isPromise(serializedValue)) {
+        throw new Error('Async-serializable type should be serialized with "deflate.resolve"');
+    }
+    return serializedValue as JsonType;
+}
+
+export namespace deflate {
+
+    /**
+     * Asynchronously serialize given instance to a JSON-compatible value.
+     * Resulting promise resolves when all nested type/property serializers are resolved
+     * @param serializable Async-serializable type instance
+     * @param options Deflate options
+     * @returns A promise which resolves to JSON-compatible type
+     */
+    export async function resolve<TOriginal>(
+        serializable: TOriginal,
+        options?: DeflateOptions<JsonType, TOriginal>
+    ): Promise<JsonType> {
+        const serializedValue = FrontendFunctions(BACKEND_NAME).deflate<JsonType, TOriginal>(serializable, options);
+        return Promise.resolve(serializedValue);
+    }
+
 }
 
 /**
- * Construct/deserialize a serializable type instance from a JSON-compatible object
+ * Construct/deserialize an instance from a JSON-compatible object
  * @param ctor Serializable type constructor function
  * @param serialized JSON-compatible object (e.g. returned from `JSON.parse`)
  * @param options Inflate options
@@ -44,7 +67,32 @@ export function inflate<TOriginal>(
     serialized: JsonType,
     options?: InflateOptions<JsonType, TOriginal>
 ): TOriginal {
-    return FrontendFunctions(BACKEND_NAME).inflate<JsonType, TOriginal>(ctor, serialized, options);
+    const originalValue = FrontendFunctions(BACKEND_NAME).inflate<JsonType, TOriginal>(ctor, serialized, options);
+    if (Util.isPromise(originalValue)) {
+        throw new Error('Async-serializable type should be deserialized with "inflate.resolve"');
+    }
+    return originalValue as TOriginal;
+}
+
+export namespace inflate {
+
+    /**
+     * Asynchronously construct/deserialize an instance from a JSON-compatible object.
+     * Resulting promise resolves when all nested type/property serializers are resolved
+     * @param ctor Async-serializable type constructor function
+     * @param serialized JSON-compatible object (e.g. returned from `JSON.parse`)
+     * @param options Inflate options
+     * @returns A promise which resolves to serializable type instance
+     */
+    export async function resolve<TOriginal>(
+        ctor: Constructor<TOriginal>,
+        serialized: JsonType,
+        options?: InflateOptions<JsonType, TOriginal>
+    ): Promise<TOriginal> {
+        const originalValue = FrontendFunctions(BACKEND_NAME).inflate<JsonType, TOriginal>(ctor, serialized, options);
+        return Promise.resolve(originalValue);
+    }
+
 }
 
 // Types