+++ lib/src/property_serializer.ts
@@ -10,10 +10,18 @@ interface PropertySerializer<TSerialized, TOriginal, TTag> {
     readonly propertyTag: TTag;
 
     /** Serializes target property from `serializable` and writes value to `serialized` */
-    down(serializable: TOriginal, serialized: TSerialized, options?: DeflateOrInflateOptions<TSerialized, TOriginal>): void;
+    down(
+        serializable: TOriginal,
+        serialized: TSerialized,
+        options?: DeflateOrInflateOptions<TSerialized, TOriginal>
+    ): void | Promise<void>;
 
     /** Deserializes target property from `serialized` and writes value to `serializable` */
-    up(serializable: TOriginal, serialized: TSerialized, options?: DeflateOrInflateOptions<TSerialized, TOriginal>): void;
+    up(
+        serializable: TOriginal,
+        serialized: TSerialized,
+        options?: DeflateOrInflateOptions<TSerialized, TOriginal>
+    ): void | Promise<void>;
 
 }